Remote container full functionality for unit test, transport bridging, ocpiremote utility -- Plus small improvements to altera scripts for programming tools and xilinx consistency
Move files into the new remote-support directory, analogous to hdl-support and ocl-support.
These files are not "driver library" files and so do not belong in the hdl/ocl/remote dirs.
We currently have no rcc-support functionality.
runtime/application/include/OcpiServer.h -> runtime/remote-support/include/OcpiServer.h
runtime/application/src/OcpiServer.cxx -> runtime/remote-support/src/OcpiServer.cxx
runtime/application/src/RemoteServer.cxx -> runtime/remote-support/src/RemoteServer.cxx
runtime/application/src/ocpiserve_main.cxx -> runtime/remote-support/src/ocpiserve_main.cxx
tools/cdk/include/hdl/altera.mk: removed from here to be more consistent with xilinx
runtime/hdl-support/altera/altera.mk: here is the new location, plus better defaults.
runtime/hdl-support/scripts/probeJtag: pass arg to altera probe script to enable help
Makefile: add remote-support dir, like hdl-support and ocl-support
Project.exports: remote-support and export new altera.mk location
components/bias_spcm.rcc/bias_spcm.cc: comment out debug code
hdl/platforms/alst4/jtagSupport_alst4: update for programming-tools-only install
-- and make more consistent with xilinx
runtime/hdl-support/altera/probeJtag_altera: ditto
os/linux/src/OcpiOsEther.cxx: add debug messages
runtime/application/src/OcpiApplication.cxx: tweak for transport bridging
-- m_scale overloaded a bit so that a scale of 1 enables transport bridging
runtime/container/include/Container.h:  comment out unused methods
runtime/container/include/ContainerBasicPort.h: update internal interfaces for transport bridging
runtime/container/include/ContainerLauncher.h: ditto
runtime/container/include/ContainerLocalPort.h: ditto
runtime/container/include/ContainerPort.h: ditto
runtime/container/include/OcpiContainerApi.h: add remote ACI functions
runtime/container/src/Container.cxx: ensure baseContainer started for bridge processing
runtime/container/src/ContainerBasicPort.cxx: transport bridging and hdl sim bridging
-- add m_address endpoint field to descriptors (was not set correctly in some places)
-- allow applyConnection and connectLocal to be used for transport bridging
runtime/container/src/ContainerLauncher.cxx: add state for transport bridging
runtime/container/src/ContainerLocalPort.cxx: the guts of transport bridging:
-- reuse the existing fancier scalability bridge ports for transport bridging
runtime/container/src/ContainerManager.cxx: remote dead code
runtime/container/src/ContainerPort.cxx: ditto, and allow explicit direction for bridge ports
runtime/container/src/ContainerWorker.cxx: transport bridging tweak
runtime/container/src/RemoteClient.cxx: adjust internal API for ocpiremote utility
runtime/dataplane/transport/impl/src/OcpiBuffer.cxx: more rigor in member initialization
runtime/dataplane/transport/impl/src/OcpiOutputBuffer.cxx: ditto
runtime/dataplane/transport/impl/src/OcpiCircuit.cxx: debug messages
runtime/dataplane/transport/impl/src/OcpiPort.cxx: fix some resource leaks for remote containers
runtime/dataplane/transport/impl/src/OcpiTransport.cxx:  ditto
runtime/dataplane/transport/impl/src/OcpiPortMetaData.cxx: simplify Xfer manager interfaces
runtime/dataplane/transport/impl/src/OcpiPortSet.cxx: debug messages
runtime/dataplane/xfer/base/include/XferEndPoint.h: remove unused args
runtime/dataplane/xfer/base/include/XferFactory.h: remove unused args and simplify interfaces
runtime/dataplane/xfer/base/include/XferManager.h: ditto
runtime/dataplane/xfer/base/include/XferServices.h: remove dead code
runtime/dataplane/xfer/base/src/XferEndPoint.cxx: fix nasty overloading of "offset zero" bug.
-- a zero offset allocation could actually happen, but zero was used as sentinel!
runtime/dataplane/xfer/base/src/XferFactory.cxx: fix resource leak for remote server, remove dead
runtime/dataplane/xfer/base/src/XferManager.cxx: remove dead code, simplify an interface
runtime/dataplane/xfer/drivers/dma/src/DtDmaXfer.cxx: more rigorous member initialization
runtime/dataplane/xfer/drivers/socket/src/DtSocketXfer.cxx: eliminate DNS dependency
- choose IP addr of first interface: connected, up, not-loopback, has MAC address, has IOP addr
runtime/dataplane/xfer/tests/src/transportTest_main.cxx: warning
runtime/hdl-support/src/ocpihdl_main.cxx: improve "ocpihdl ethers" printout, includ iface #s
runtime/hdl/include/HdlContainer.h: remove custom container stop()/start() - unneeded.
runtime/hdl/src/HdlContainer.cxx: fix socket system id, use better "addTransport" base method
runtime/hdl/src/HdlLSimDevice.cxx: fix to make implicit transport bridging work
-- it should have always worked by didn't
-- also enable multiple simulators again, hopefully we can keep this open
runtime/rcc/include/RccContainer.h: remove unused methods
runtime/rcc/src/RccContainer.cxx: ditto
runtime/remote-support/include/OcpiServer.h: record whether containers need transport bridging
runtime/remote-support/src/OcpiServer.cxx: ditto
runtime/remote-support/src/RemoteServer.cxx: ditto
runtime/remote/include/RemoteServer.h: ditto
runtime/remote-support/src/ocpiserve_main.cxx: properly implement "remove artifact cache" option
runtime/remote/include/RemoteClient.h: change internal interface for ocpiremote
runtime/remote/include/RemoteDriver.h: move class Driver definition into this file (be normal)
runtime/remote/src/RemoteDriver.cxx: add probe printing like other ocpixxx functions, reindent
runtime/util/assembly/src/OcpiUtilAssembly.cxx: allow unsetting/overriding default assignments
runtime/util/res/src/OcpiAlloc.cxx: remove debug
tools/cdk/scripts/util.sh: add OcpiEcho for -n
tools/ocpigen/src/ocpigen_main.cxx: suppress discovery for this tool of course
tools/ocpigen/src/tests.cxx: improve quoting for property values that are "<none>"
runtime/remote-support/Makefile.ocpi.imports: support for new dir
runtime/remote-support/src/ocpiremote_main.cxx: new utility for remote stuff

+++ runtime/remote-support/src/ocpiserve_main.cxx
@@ -78,7 +78,8 @@ static void
 sigint(int /* signal */) {
   signal(SIGINT, SIG_DFL);
   signal(SIGTERM, SIG_DFL);
-  rmdir();
+  if (options.remove())
+    rmdir();
   options.bad("interrupted");
 }
 
@@ -89,10 +90,10 @@ static int mymain(const char **) {
   OCPI::OS::logSetLevel(options.loglevel());
   setenv("OCPI_LIBRARY_PATH", options.directory(), 1);
   // Catch signals in order to delete the artifact cache
-  if (options.remove()) {
-    ocpiCheck(signal(SIGINT, sigint) != SIG_ERR);
+  // Catch SIGINT all the time to make valgrind happe
+  ocpiCheck(signal(SIGINT, sigint) != SIG_ERR);
+  if (options.remove())
     ocpiCheck(signal(SIGTERM, sigint) != SIG_ERR);
-  }
   OCPI::Driver::ManagerManager::configure();
   assert(OCPI::Library::Library::s_firstLibrary);
   for (unsigned n = 1; n < options.processors(); n++) {
